// WADNR.Web/src/app/shared/generated/model/project-external-link-save-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectExternalLinkItemRequest } from './project-external-link-item-request';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectExternalLinkSaveRequest { 
    ExternalLinks?: Array<ProjectExternalLinkItemRequest> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectExternalLinkSaveRequestForm { 
    ExternalLinks?: FormControl<Array<ProjectExternalLinkItemRequest>>;
}

export class ProjectExternalLinkSaveRequestFormControls { 
    public static ExternalLinks = (value: FormControlState<Array<ProjectExternalLinkItemRequest>> | Array<ProjectExternalLinkItemRequest> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectExternalLinkItemRequest>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}