// WADNR.Web/src/app/shared/generated/model/project-external-link-item-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectExternalLinkItemRequest { 
    ProjectExternalLinkID?: number | null;
    ExternalLinkLabel: string;
    ExternalLinkUrl: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectExternalLinkItemRequestForm { 
    ProjectExternalLinkID?: FormControl<number>;
    ExternalLinkLabel: FormControl<string>;
    ExternalLinkUrl: FormControl<string>;
}

export class ProjectExternalLinkItemRequestFormControls { 
    public static ProjectExternalLinkID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ExternalLinkLabel = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(1),
                Validators.maxLength(300),
            ],
        }
    );
    public static ExternalLinkUrl = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(1),
                Validators.maxLength(300),
            ],
        }
    );
}